+++ front/src/app/idea-memos/(index)/IdeaMemos.module.scss
