+++ front/src/app/idea-memos/(index)/IdeaMemosPresentation.tsx
