+++ front/src/app/idea-memos/(index)/page.tsx
@@ -1,4 +1,4 @@
-import IdeaMemosPresentation from "@/app/idea-memos/IdeaMemosPresentation";
+import IdeaMemosPresentation from "@/app/idea-memos/(index)/IdeaMemosPresentation";
 import { getAllIdeaMemos } from "@/lib/idea-memos";
 import { IdeaMemoType } from "@/types";
