+++ front/src/app/idea-memos/(show)/[uuid]/page.tsx
@@ -1,7 +1,7 @@
-import IdeaMemosUuidPresentation from "@/app/presentation/IdeaMemosUuid/IdeaMemosUuidPresentation";
+import IdeaMemosUuidContainer from "@/app/idea-memos/(show)/[uuid]/IdeaMemosUuidContainer";
 import { getIdeaMemo } from "@/lib/idea-memos";
 
 export default async function page({ params }: { params: { uuid: string } }) {
   const ideaMemo = await getIdeaMemo(params.uuid);
-  return <IdeaMemosUuidPresentation ideaMemo={ideaMemo} />;
+  return <IdeaMemosUuidContainer ideaMemo={ideaMemo} />;
 }
