--- front/src/app/presentation/IdeaMemosUuid/IdeaMemosUuidPresentation.module.scss
@@ -1,178 +1,0 @@
-.wrapper {
-  min-height: calc(100vh - var(--header-height));
-  width: 100vw;
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  height: calc(100% - var(--back-button-length) - var(--page-padding));
-  width: calc(100% - var(--page-padding) * 2);
-  padding-bottom: calc(var(--back-button-length) + var(--page-padding) + 80px);
-
-  @media screen and (min-width: 1024px) {
-    height: calc(100% - var(--back-button-length) - var(--page-padding-lg));
-    width: calc(100% - var(--page-padding-lg) * 2);
-    padding-bottom: calc(
-      var(--back-button-length) + var(--page-padding-lg) + 160px
-    );
-  }
-}
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  @media screen and (min-width: 768px) {
-    width: 60%;
-  }
-}
-
-.cardContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-}
-
-/* header */
-.cardHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  padding: 12px 16px 16px 0;
-
-  @media screen and (min-width: 768px) {
-    padding: 16px 20px 20px 0;
-  }
-}
-
-// .heartIcon {
-//   height: 40px;
-//   width: auto;
-// }
-
-/* body */
-.cardBody {
-  width: 100%;
-  padding: 0 16px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 24px;
-
-  @media screen and (min-width: 768px) {
-    padding: 0 20px;
-  }
-}
-
-.form {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 16px;
-}
-
-.save {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.section {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 16px;
-  width: 100%;
-}
-
-.sectionContentContainer {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  width: 100%;
-}
-
-.sectionContent {
-  width: 98%;
-}
-
-/* footer */
-.cardFooter {
-  width: 100%;
-  text-align: end;
-  padding: 16px;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  position: relative;
-
-  @media screen and (min-width: 768px) {
-    padding: 20px;
-  }
-}
-
-/* 編集・削除・キャンセルボタン */
-.buttonContainer {
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-top: 32px;
-}
-
-.buttons {
-  display: flex;
-  justify-content: canter;
-  align-items: center;
-  gap: 8px;
-}
-
-.button {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-}
-
-.cancelButton {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-}
-
-/* エラー */
-.error {
-  color: var(--red);
-  font-size: 16px;
-  margin: 8px 0 4px;
-}
-
-/* 成功メッセージ */
-.message {
-  color: var(--light-blue);
-  font-size: 14px;
-  margin: 8px 0;
-  width: 100%;
-  text-align: center;
-}
