+++ front/src/features/idea-memos-uuid/components/ButtonSection/ButtonSection.tsx
@@ -1,0 +1,13 @@
+import styles from "@/features/idea-memos-uuid/components/ButtonSection/ButtonSection.module.scss";
+
+export default function ButtonSection({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className={styles.buttonContainer}>
+      <div className={styles.buttons}>{children}</div>
+    </div>
+  );
+}
