+++ front/src/features/idea-memos-uuid/components/CancelButton/CancelButton.module.scss
@@ -1,0 +1,15 @@
+.buttonContainer {
+  width: 100%;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  margin-top: 32px;
+}
+
+.cancelButton {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+}
