+++ front/src/features/idea-memos-uuid/components/CardBodyOnViewMode/CardBodyOnViewMode.module.scss
@@ -1,0 +1,33 @@
+.cardBody {
+  width: 100%;
+  padding: 0 16px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 24px;
+
+  @media screen and (min-width: 768px) {
+    padding: 0 20px;
+  }
+}
+
+.section {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 16px;
+  width: 100%;
+}
+
+.sectionContentContainer {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  width: 100%;
+}
+
+.sectionContent {
+  width: 98%;
+}
