+++ front/src/features/idea-memos-uuid/components/DeleteButton/DeleteButton.module.scss
@@ -1,0 +1,8 @@
+.button {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+}
