+++ front/src/features/idea-memos-uuid/components/IdeaMemoCardHeader/IdeaMemoCardHeader.tsx
@@ -1,0 +1,9 @@
+import styles from "@/features/idea-memos-uuid/components/IdeaMemoCardHeader/IdeaMemoCardHeader.module.scss";
+
+export default function IdeaMemoCardHeader() {
+  return (
+    <div className={styles.cardHeader}>
+      {/* <GoHeart className={styles.heartIcon} /> */}
+    </div>
+  );
+}
